+++ app/src/main/java/com/example/kgamify/profilepage.java
@@ -88,13 +88,18 @@ public class profilepage extends AppCompatActivity
 
 
                     case R.id.logout:
-                        SharedPreferences.Editor editor=sharedPreferences.edit();
-                        editor.clear();
-                        editor.commit();
-                        finish();
-                        Toast.makeText(getApplicationContext(),"Log out successfully!!",Toast.LENGTH_SHORT).show();
-                        Intent i=new Intent(getApplicationContext(),MainActivity.class);
-                        startActivity(i);
+                        if(current_user_phone!=null){
+                            SharedPreferences.Editor editor=sharedPreferences.edit();
+                            editor.clear();
+                            editor.commit();
+                            finish();
+                            Toast.makeText(getApplicationContext(),"Log out successfully!!",Toast.LENGTH_SHORT).show();
+                            Intent i=new Intent(getApplicationContext(),MainActivity.class);
+                            startActivity(i);
+                        }
+                        else {
+                            Toast.makeText(getApplicationContext(),"You are not Logged In",Toast.LENGTH_SHORT).show();
+                        }
                         break;
                 }
                 return true;
